Fix Bugs in Updates Section
