+++ lib/presentation/widgets/updates/update_item_builder.dart
@@ -1,6 +1,8 @@
 import 'package:college_cupid/domain/models/update_model.dart';
 import 'package:college_cupid/presentation/widgets/updates/match_update_card.dart';
-import 'package:college_cupid/presentation/widgets/updates/standard_update_card.dart';
+import 'package:college_cupid/presentation/widgets/updates/confession_reply_card.dart';
+import 'package:college_cupid/presentation/widgets/updates/voice_note_reply_card.dart';
+import 'package:college_cupid/presentation/widgets/updates/profile_reply_card.dart';
 import 'package:flutter/material.dart';
 
 class UpdateItemBuilder extends StatelessWidget {
@@ -14,9 +16,12 @@ class UpdateItemBuilder extends StatelessWidget {
       case UpdateType.match:
         return MatchUpdateCard(update: update);
       case UpdateType.voiceReply:
+        return VoiceNoteReplyCard(update: update);
+      case UpdateType.profileReply:
+        return ProfileReplyCard(update: update);
       case UpdateType.textReply:
       case UpdateType.confessionReply:
-        return StandardUpdateCard(update: update);
+        return ConfessionReplyCard(update: update);
     }
   }
 }
